Remove test file and update leaderboard fetch
